perf: extract css
