+++ .gitignore
@@ -14,6 +14,7 @@ tests/clover*.xml
 node_modules/
 
 /js
+/css
 
 # packaged app
 build/artifacts
